Move to flat config
