+++ .npmignore
@@ -1,10 +1,11 @@
-node_modules
-.eslintrc.js
 .editorconfig
-pnpm-lock.yaml
-rules/
-test.js
-.vscode/
-test/
+.eslintrc.js
 .github/
 .ncurc
+.vscode/
+eslint.config.js
+find/
+node_modules
+pnpm-lock.yaml
+sort/
+test/
