+++ jsdoc_ts.js
@@ -1,0 +1,12 @@
+'use strict';
+
+module.exports = {
+  files: [
+    '**/*.ts',
+  ],
+  rules: {
+    'jsdoc/no-undefined-types': 'off', // Switch to typedoc
+    'jsdoc/require-param-type': 'off', // Not needed in TS
+    'jsdoc/require-returns-type': 'off', // Not needed in TS
+  },
+};
