+++ mjs.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = {
+  files: [
+    '**/*.mjs',
+    '**/*.ts',
+  ],
+  languageOptions: {
+    sourceType: 'module',
+  },
+};
