+++ module.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  files: ['**/*.js'],
+  languageOptions: {
+    sourceType: 'module',
+  },
+};
