+++ sort-rules/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "eslint-plugin-hildjj",
+  "name": "@hildjj/sort-rules",
   "version": "0.0.0",
   "description": "eslint plugins for eslint config",
   "private": true,
@@ -11,15 +11,17 @@
     "directory": "rules"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "6.19.1",
-    "@typescript-eslint/parser": "6.19.1",
-    "eslint": "^8.56.0",
+    "@stylistic/eslint-plugin": "1.6.3",
+    "@typescript-eslint/eslint-plugin": "7.1.1",
+    "@typescript-eslint/parser": "7.1.1",
+    "eslint": "^8.57.0",
     "eslint-plugin-ava": "^14.0.0",
-    "eslint-plugin-jsdoc": "^48.0.2",
+    "eslint-plugin-jsdoc": "^48.2.1",
     "eslint-plugin-n": "16.6.2",
-    "typescript": "5.3.3"
+    "typescript": "5.4.2"
   },
+  "packageManager": "pnpm@8.15.4",
   "engines": {
-    "node": ">=16"
+    "node": ">=18"
   }
 }
