+++ test/eslint.config.js
@@ -1,0 +1,15 @@
+import ava from '@cto.af/eslint-config/ava.js';
+import js from '@cto.af/eslint-config/js.js';
+import jsdoc from '@cto.af/eslint-config/jsdoc.js';
+import jsdoc_ts from '@cto.af/eslint-config/jsdoc_ts.js';
+import mod from '@cto.af/eslint-config/module.js';
+import ts from '@cto.af/eslint-config/ts.js';
+
+export default [
+  js,
+  mod,
+  ts,
+  ava,
+  jsdoc,
+  jsdoc_ts,
+];
